Update media queries to only appy to webkit browsers
Since zoom is not supported across all browsers, the media query that relies on it has been modified to only apply
to webkit browsers (which support it). A more permanent solution that will optimize the application across browsers
and not use zoom will most likely replace this in the future.